requirements.txt
boto3
pandas
requests
python-dotenv
pyarrow